backend/requirements.txt
# ========================
# Django & REST
# ========================
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt>=5.2.0

# ========================
# Celery + Redis
# ========================
celery[redis]==5.3.6
redis==5.0.3
django-celery-results==2.5.1

# ========================
# AWS S3 Integration
# ========================
boto3==1.34.67
django-storages==1.14.2

# ========================
# Translation & NLP
# ========================
deep-translator==1.11.4
transformers==4.43.3
torch==2.2.0

# ========================
# OpenAI / LangChain
# ========================
openai==1.12.0
langchain==0.1.16
langsmith==0.1.17
langgraph-sdk==0.2.9


# ========================
# HTTP Client
# ========================
httpx==0.27.0

# ========================
# Video Processing
# ========================
# moviepy==2.2.1
# imageio==2.37.0
# imageio-ffmpeg==0.6.0
# proglog==0.1.12
decorator==5.2.1
numpy==1.26.4
Pillow==10.4.0
ffmpeg-python==0.2.0
opencv-python==4.10.0.84


# ========================
# Whisper
# ========================
whisper==1.1.10
soundfile==0.12.1

# ========================
# Misc
# ========================
python-dotenv==1.0.1
tqdm==4.67.1

psycopg2-binary
